fix: clean up raduino menu

+++ tools/scripts/podman.sh
@@ -6,7 +6,10 @@ if [ "$1" == "local" ]
 then
     podman image exists docker-daemon:radio-arduino:${TAG} && podman pull docker-daemon:radio-arduino:${TAG}
     podman run --network host -it --rm --device=/dev/ttyUSB0 radio-arduino:${TAG} bash
-else
+fi
+
+if [ "$1" == "remote" ]
+then
     podman run --network host -it --rm --device=/dev/ttyUSB0 ghcr.io/larsindresaeter/radio-arduino:latest bash
 fi
